Make preload play nice with CSP strict-dynamic
Dynamically loaded preloaded scripts were not loaded when a strict-dynamic
CSP directive was used. This was because the state regarding whether the
element was parser created or not was not properly propoagated to the CSP
checks. This CL fixes that and adds tests that make sure that such scripts
can be preloaded (and that markup based preloaded scripts with no nonce cannot).

Bug: 752922
Change-Id: Ib4f3bce8583ccc9770c261d76167243033e5a112